backend/requirements.txt
fastapi==0.135.1
uvicorn==0.30.0
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
groq==0.1.1
